gvastreammux: move meta files to shared metadata library
